// src/app/_dto/saldetail.dto.ts
import { empCodedD } from "./empCodeD.dto";

export class apsaldetailDto{
    coCode: string;

    dvCode: string;
    
    brCode: string;
    
    intCode: string;
    
    vrNo: number;
    
    srNo: number;
    
    vrDate: Date;
    
    attYear: number;
    
    attMnth: number;
    
    attStartDate: Date;
    
    empCode: string;
    
    empCat: string;
    
    salCat: string;
    
    salPostCat: string;
    
    empDept: string;
    
    empBr: string;
    
    salType: string;
    
    payMode: string;
    
    paidYn: string;
    
    paidDt: Date;
    
    payableDt: Date;
    
    paidAmt: number;
    
    remarks: string;
    
    presentDays: number;
    
    absentDays: number;
    
    sickDays: number;
    
    emergencyDays: number;
    
    leaveWoPayDays: number;
    
    annualLeaveDays: number;
    
    examDays: number;
    
    maternityDays: number;
    
    death1Days: number;
    
    death2Days: number;
    
    husDeathDays: number;
    
    hajDays: number;
    
    marriageDays: number;
    
    ot1Hrs: number;
    
    ot2Hrs: number;
    
    ot3Hrs: number;
    
    ot4Hrs: number;
    
    a1: number;
    
    a2: number;
    
    a3: number;
    
    a4: number;
    
    a5: number;
    
    a6: number;
    
    a7: number;
    
    a8: number;
    
    a9: number;
    
    a10: number;
    
    a11: number;
    
    a12: number;
    
    a13: number;
    
    a14: number;
    
    a15: number;
    
    a16: number;
    
    a17: number;
    
    a18: number;
    
    a19: number;
    
    a20: number;
    
    a21: number;
    
    a22: number;
    
    a23: number;
    
    a24: number;
    
    a25: number;
    
    a26: number;
    
    a27: number;
    
    a28: number;
    
    a29: number;
    
    a30: number;
    
    aGross: number;    
    
    aOth: number;
    
    aOt1: number;
    
    aOt2: number;
    
    aOt3: number;
    
    aOt4: number;
    
    d1: number;
    
    d2: number;
    
    d3: number;
    
    d4: number;
    
    d5: number;
    
    d6: number;
    
    d7: number;
    
    d8: number;
    
    d9: number;
    
    d10: number;
    
    d11: number;
    
    d12: number;
    
    d13: number;
    
    d14: number;
    
    d15: number;
    
    d16: number;
    
    d17: number;
    
    d18: number;
    
    d19: number;
    
    d20: number;
    
    dGross: number;
    
    dOth: number;
    
    dAdv: number;
    
    dLoan: number;
    
    dSick: number;
    
    dPasi: number;
    
    a1Now: number;
    
    a2Now: number;
    
    a3Now: number;
    
    a4Now: number;
        
    a5Now: number;
    
    a6Now: number;
    
    a7Now: number;
    
    a8Now: number;
    
    a9Now: number;
    
    a10Now: number;
    
    a11Now: number;
    
    a12Now: number;
    
    a13Now: number;
    
    a14Now: number;
    
    a15Now: number;
    
    a16Now: number;
    
    a17Now: number;
    
    a18Now: number;
    
    a19Now: number;
    
    a20Now: number;
    
    a21Now: number;
    
    a22Now: number;
    
    a23Now: number;
    
    a24Now: number;
    
    a25Now: number;
    
    a26Now: number;
    
    a27Now: number;
    
    a28Now: number;
    
    a29Now: number;
    
    a30Now: number;
    
    aGrossNow: number;
    
    refCo: string;
    
    refDv: string;
    
    refBr: string;
    
    refInt: string;
    
    refVrNo: number;
    
    refSrNo: number;
    
    verified: string;
    
    verifiedDate: Date;
    
    verifiedUser: string;
    
    cancel: string;
    
    voiddate: Date;
    
    voiduser: string;
    
    sysuser: string;
    
    systemdate: Date;
    
    modiuser: string;
    
    modidate: Date;
    
    mnthDays: number;
    
    empName?: string;

    empCodeD?:empCodedD[];

}